voila

+++ paysan-service/src/main/java/org/gov/moussaada/paysan_service/dao/AddresseDAO.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public interface AddresseDAO extends JpaRepository<Addresse,Integer> {
 
     @Query("FROM Addresse a WHERE a.id_utilisateur = :id")
-    Optional<Addresse> findByUser(Long id);
+    Optional<Addresse> findByUser(int id);
 
     @Query("FROM Ville v")
     List<Ville> getAllVille();
